requirements.txt
biopython==1.85
numpy==2.2.2